backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.6
djangorestframework-simplejwt>=5.2.0
python-decouple>=3.8
drf-spectacular>=0.26.0
gunicorn>=21.0.0
whitenoise>=6.5.0
Pillow>=10.0.0
django-filter>=23.0.0
